fix some problems for the tasks
