+++ paddlenlp/taskflow/lexical_analysis.py
@@ -84,14 +84,12 @@ class LacTask(Task):
     Args:
         task(string): The name of task.
         model(string): The model name in the task.
-        static_mode(bool): The flag to control in the static/dygraph mode.
         kwargs (dict, optional): Additional keyword arguments passed along to the specific task. 
     """
 
-    def __init__(self, task, model, static_mode, **kwargs):
-        super().__init__(
-            task=task, model=model, static_mode=static_mode, **kwargs)
-        self.static_mode = False
+    def __init__(self, task, model, **kwargs):
+        super().__init__(task=task, model=model, **kwargs)
+        self._static_mode = False
         self._usage = usage
         word_dict_path = download_file(
             self._task_path, "lac_params" + os.path.sep + "word.dic",
@@ -109,7 +107,7 @@ class LacTask(Task):
             zip(self._word_vocab.values(), self._word_vocab.keys()))
         self._id2tag_dict = dict(
             zip(self._tag_vocab.values(), self._tag_vocab.keys()))
-        if self.static_mode:
+        if self._static_mode:
             self._get_inference_model()
         else:
             self._construct_model(model)
@@ -204,7 +202,7 @@ class LacTask(Task):
         """
         results = []
         lens = []
-        if not self.static_mode:
+        if not self._static_mode:
             with dygraph_mode_guard():
                 for batch in inputs['data_loader']:
                     input_ids, seq_len = batch
